修改各个实体中  基因网中的实体主键类型 由Long更改为Integer

+++ src/main/java/com/agen/orderdiscount/entity/Discount.java
@@ -36,12 +36,12 @@ public class Discount {
      * 机构ID
      */
     @Column(nullable = false)
-    private Long adminId;
+    private Integer adminId;
     /**
      * 产品ID
      */
     @Column(nullable = false)
-    private Long productId;
+    private Integer productId;
     /**
      * 产品价格
      */
